Revert "Preserve Codex [tui] settings across managed CODEX_HOME remirrors (#9475)"
This reverts commit c8381f3ea780d9f0da9aa58d88d515577ca3e399.
