+++ src/main/codex/config-settings-promotion.ts
@@ -18,8 +18,6 @@ import {
   isTomlStructuralLine,
   updateTomlLineScanState
 } from './config-toml-line-scan'
-import { parseTomlKeyPath, parseTomlTableHeaderPath } from './config-toml-key-path'
-import { tuiStructuredKey, upsertPromotedSettingsInContent } from './codex-config-settings-upsert'
 
 // Why: the mirror reverts in-Codex config changes each launch; promotion salvages them by diffing the last baseline.
 
@@ -30,45 +28,6 @@ export const PROMOTED_CODEX_SETTING_KEYS = [
   'approval_policy',
   'sandbox_mode'
 ] as const
-
-// Why: the [tui] keys the Codex TUI's user-facing pickers persist (status line,
-// terminal title, theme). Like the top-level list, every key here gets written
-// into the user's real ~/.codex/config.toml on promotion — grow it deliberately.
-export const PROMOTED_CODEX_TUI_SETTING_KEYS = [
-  'status_line',
-  'status_line_use_colors',
-  'terminal_title',
-  'theme'
-] as const
-
-// Why: promotion diffs and upserts operate on structured keys — top-level keys
-// keep their bare name, [tui] keys are namespaced tui.<key> so their baseline
-// entries cannot collide with a top-level key of the same name.
-const PROMOTED_STRUCTURED_KEYS: readonly string[] = [
-  ...PROMOTED_CODEX_SETTING_KEYS,
-  ...PROMOTED_CODEX_TUI_SETTING_KEYS.map(tuiStructuredKey)
-]
-
-function isPromotedTuiKey(key: string): boolean {
-  return (PROMOTED_CODEX_TUI_SETTING_KEYS as readonly string[]).includes(key)
-}
-
-// Returns the structured tui key a scanned line's key represents, or null. In
-// the preamble it recognizes the dotted `tui.<key>` form a user may hand-author;
-// inside the first `[tui]` table body it recognizes the bare `<key>` form Codex
-// writes. Both map to the same structured key so either config shape promotes.
-function matchTuiStructuredKey(
-  keyPath: string[],
-  inPreamble: boolean,
-  tuiBodyActive: boolean
-): string | null {
-  if (inPreamble) {
-    const tuiKey = keyPath.length === 2 && keyPath[0] === 'tui' ? keyPath[1] : null
-    return tuiKey && isPromotedTuiKey(tuiKey) ? tuiStructuredKey(tuiKey) : null
-  }
-  const tuiKey = keyPath.length === 1 ? keyPath[0] : null
-  return tuiBodyActive && tuiKey && isPromotedTuiKey(tuiKey) ? tuiStructuredKey(tuiKey) : null
-}
 
 type TopLevelSettingValue = {
   raw: string
@@ -111,67 +70,24 @@ function readSettingsBaseline(runtimeHomePath: string): Map<string, string> | nu
   }
 }
 
-function matchPromotedStructuredKey(
-  line: string,
-  inPreamble: boolean,
-  tuiBodyActive: boolean
-): { structuredKey: string; raw: string } | null {
-  const parsed = parseTomlKeyPath(line)
-  if (!parsed || line[parsed.end] !== '=') {
-    return null
-  }
-  const raw = line.slice(parsed.end + 1).trim()
-  const topLevelKey = parsed.segments.length === 1 ? parsed.segments[0] : null
-  if (
-    inPreamble &&
-    topLevelKey &&
-    (PROMOTED_CODEX_SETTING_KEYS as readonly string[]).includes(topLevelKey)
-  ) {
-    return { structuredKey: topLevelKey, raw }
-  }
-  const tuiKey = matchTuiStructuredKey(parsed.segments, inPreamble, tuiBodyActive)
-  return tuiKey ? { structuredKey: tuiKey, raw } : null
-}
-
-// Why: top-level preamble scalars keep the historical behavior; [tui] keys are
-// collected from the first bare [tui] table body or the dotted preamble form,
-// keyed by structured path. Any table header (including [tui.*] subtables) ends
-// the [tui] body, and [profiles.*]/other tables are still ignored.
-function readPromotedSettingValues(configPath: string): Map<string, TopLevelSettingValue> {
+// Why: only top-level preamble keys are scanned; rewriting nested [profiles.*] tables isn't worth the risk here.
+function readTopLevelSettingValues(configPath: string): Map<string, TopLevelSettingValue> {
   const result = new Map<string, TopLevelSettingValue>()
   if (!existsSync(configPath)) {
     return result
   }
   const lines = readFileSync(configPath, 'utf-8').split('\n')
   let state = createTomlLineScanState()
-  let inPreamble = true
-  let tuiTableSeen = false
-  let tuiBodyActive = false
   for (const line of lines) {
     if (isTomlStructuralLine(state)) {
-      const header = getTomlTableHeader(line)
-      if (header) {
-        const table = parseTomlTableHeaderPath(header)
-        tuiBodyActive =
-          table !== null &&
-          !table.isArray &&
-          table.segments.length === 1 &&
-          table.segments[0] === 'tui' &&
-          !tuiTableSeen
-        if (tuiBodyActive) {
-          tuiTableSeen = true
-        }
-        inPreamble = false
-        state = updateTomlLineScanState(state, line)
-        continue
+      if (getTomlTableHeader(line)) {
+        break
       }
-      const matched = matchPromotedStructuredKey(line, inPreamble, tuiBodyActive)
-      if (matched) {
+      const match = /^[ \t]*([A-Za-z0-9_-]+)[ \t]*=[ \t]*(.*?)[ \t\r]*$/.exec(line)
+      const key = match?.[1]
+      if (key && (PROMOTED_CODEX_SETTING_KEYS as readonly string[]).includes(key)) {
         const nextState = updateTomlLineScanState(state, line)
-        result.set(matched.structuredKey, {
-          raw: matched.raw,
-          multiline: !isTomlStructuralLine(nextState)
-        })
+        result.set(key, { raw: match?.[2] ?? '', multiline: !isTomlStructuralLine(nextState) })
         state = nextState
         continue
       }
@@ -193,7 +109,7 @@ export function snapshotCodexRuntimeSettingsBaseline(
     const runtimeTomlPath = join(runtimeHomePath, 'config.toml')
     // Why: record an empty baseline even for a missing runtime config, so Codex's first write still diffs and promotes.
     const settings: Record<string, string> = {}
-    for (const [key, value] of readPromotedSettingValues(runtimeTomlPath)) {
+    for (const [key, value] of readTopLevelSettingValues(runtimeTomlPath)) {
       if (!value.multiline) {
         settings[key] = value.raw
       }
@@ -257,9 +173,10 @@ function promoteCodexRuntimeSettingsToSystemUnsafe(homes: CodexSettingsPromotion
   if (!baseline) {
     return
   }
-  const runtimeValues = readPromotedSettingValues(runtimeTomlPath)
-  const changedRuntimeValues = new Map<string, string>()
-  for (const key of PROMOTED_STRUCTURED_KEYS) {
+  const runtimeValues = readTopLevelSettingValues(runtimeTomlPath)
+  const systemValues = readTopLevelSettingValues(systemTomlPath)
+  const updates = new Map<string, string>()
+  for (const key of PROMOTED_CODEX_SETTING_KEYS) {
     const runtime = runtimeValues.get(key)
     if (!runtime || runtime.multiline) {
       continue
@@ -268,14 +185,6 @@ function promoteCodexRuntimeSettingsToSystemUnsafe(homes: CodexSettingsPromotion
       // Orca mirrored this value and nothing touched it since — not a change.
       continue
     }
-    changedRuntimeValues.set(key, runtime.raw)
-  }
-  if (changedRuntimeValues.size === 0) {
-    return
-  }
-  const systemValues = readPromotedSettingValues(systemTomlPath)
-  const updates = new Map<string, string>()
-  for (const [key, runtimeRaw] of changedRuntimeValues) {
     const system = systemValues.get(key)
     if (system?.multiline) {
       continue
@@ -284,7 +193,7 @@ function promoteCodexRuntimeSettingsToSystemUnsafe(homes: CodexSettingsPromotion
     if (system?.raw !== baseline.get(key)) {
       continue
     }
-    updates.set(key, runtimeRaw)
+    updates.set(key, runtime.raw)
   }
   if (updates.size === 0) {
     return
@@ -296,10 +205,7 @@ function promoteCodexRuntimeSettingsToSystemUnsafe(homes: CodexSettingsPromotion
   mkdirSync(dirname(writeTarget.path), { recursive: true, mode: 0o700 })
   const targetExists = existsSync(writeTarget.path)
   const systemContent = targetExists ? readFileSync(writeTarget.path, 'utf-8') : ''
-  const nextContent = upsertPromotedSettingsInContent(systemContent, updates)
-  if (nextContent === systemContent) {
-    return
-  }
+  const nextContent = upsertTopLevelSettingsInContent(systemContent, updates)
   if (targetExists && parseWslUncPath(writeTarget.path)) {
     // Why: \\wsl$ 9P symlink metadata is unreliable; write through the existing file to preserve the WSL-side inode.
     writeFileSync(writeTarget.path, nextContent, 'utf-8')
@@ -345,4 +251,56 @@ function resolveDanglingSymlinkTarget(linkPath: string): string {
   }
   // Why: replacing any link in a cycle would destroy dotfile-manager state; abort instead.
   throw new Error(`Codex config symlink cycle at ${linkPath}`)
+}
+
+export function upsertTopLevelSettingsInContent(
+  content: string,
+  updates: Map<string, string>
+): string {
+  const lines = content.split('\n')
+  let state = createTomlLineScanState()
+  let preambleEnd = lines.length
+  const keyLineIndexes = new Map<string, number>()
+  for (let index = 0; index < lines.length; index += 1) {
+    const line = lines[index] ?? ''
+    if (isTomlStructuralLine(state)) {
+      if (getTomlTableHeader(line)) {
+        preambleEnd = index
+        break
+      }
+      const match = /^[ \t]*([A-Za-z0-9_-]+)[ \t]*=/.exec(line)
+      if (match?.[1] && updates.has(match[1])) {
+        keyLineIndexes.set(match[1], index)
+      }
+    }
+    state = updateTomlLineScanState(state, line)
+  }
+
+  // Why: match the file's existing EOL (CRLF split leaves a trailing \r) so a Windows config doesn't go mixed-EOL.
+  const usesCrlf = content.includes('\r\n')
+  const insertions: string[] = []
+  for (const [key, raw] of updates) {
+    const existingIndex = keyLineIndexes.get(key)
+    const rendered = `${key} = ${raw}`
+    if (existingIndex !== undefined) {
+      lines[existingIndex] = lines[existingIndex]?.endsWith('\r') ? `${rendered}\r` : rendered
+    } else {
+      insertions.push(usesCrlf ? `${rendered}\r` : rendered)
+    }
+  }
+  if (insertions.length > 0) {
+    let insertAt = preambleEnd
+    while (insertAt > 0 && (lines[insertAt - 1] ?? '').trim() === '') {
+      insertAt -= 1
+    }
+    if (insertAt === preambleEnd && preambleEnd < lines.length) {
+      insertions.push(usesCrlf ? '\r' : '')
+    }
+    lines.splice(insertAt, 0, ...insertions)
+  }
+  const result = lines.join('\n')
+  if (result.endsWith('\n') || result.length === 0) {
+    return result
+  }
+  return result.endsWith('\r') ? `${result}\n` : `${result}${usesCrlf ? '\r\n' : '\n'}`
 }
